Add vertical resourceViewMode
